new_fw: Add tuning knob for pid activation
This adds a new setting 'pidSwitch' which is the distance in degrees
celcius from the target temperature at which to activate the pid
control of the wattage.  Prior to that point, it fires at initWatts.

The default for this is 600C, which maintains the original behavior.
Values less than 200 easily lead to overshoot.